requirements.txt
torch
torchvision
transformers
datasets
accelerate
pillow
peft
python-Levenshtein